add sign up and hashing

+++ sar-frontend/src/pages/SignUp.tsx
@@ -13,27 +13,35 @@ import { useNavigate } from "react-router-dom";
 import { paths } from "RootPage";
 import axios from "axios";
 
-export const LoginPage: React.FC = () => {
+export const SignUpPage: React.FC = () => {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
+  const [passwordConfirm, setPasswordConfirm] = useState("");
   const [error, setError] = useState("");
 
   const navigate = useNavigate();
 
-  const handleLogin = () => {
-    axios.post(`${apiUrl}/auth/login`, { email, password }).then((response) => {
-      if (response.status === 200) {
-        localStorage.setItem("userProfile", JSON.stringify(response.data));
-        navigate(paths.home);
-      } else {
-        setError("Invalid email or password");
-      }
-    });
+  const handleSignUp = () => {
+    if (password !== passwordConfirm) {
+      setError("Passwords do not match");
+      return;
+    }
+
+    axios
+      .post(`${apiUrl}/auth/register`, { email, password })
+      .then((response) => {
+        if (response.status === 200) {
+          localStorage.setItem("userProfile", JSON.stringify(response.data));
+          navigate(paths.home);
+        } else {
+          setError("Invalid email or password");
+        }
+      });
   };
 
   return (
     <Box>
-      <Heading>Login Page</Heading>
+      <Heading>Sign Up</Heading>
       {error && <Text color={"red"}>{error}</Text>}
       <FormControl>
         <FormLabel>Email</FormLabel>
@@ -53,8 +61,16 @@ export const LoginPage: React.FC = () => {
           onChange={(e) => setPassword(e.target.value)}
         />
       </FormControl>
-      <Button onClick={handleLogin}>Login</Button>
-      <Button onClick={() => navigate(paths.signup)}>Sign Up</Button>
+      <FormControl>
+        <FormLabel>Confirm Password</FormLabel>
+        <Input
+          type="password"
+          placeholder="Password"
+          value={passwordConfirm}
+          onChange={(e) => setPasswordConfirm(e.target.value)}
+        />
+      </FormControl>
+      <Button onClick={handleSignUp}>Sign Up</Button>
     </Box>
   );
 };
